+++ packages/react-magma-dom/src/components/List/List.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
-import styled from '../../theme/styled';
 import { css } from '@emotion/core';
 import { ThemeContext } from '../../theme/ThemeContext';
 import { magma, ThemeInterface } from '../../theme/magma';
 import { TypographyVisualStyle, TypographyComponent } from '../Typography';
 import { InverseContext, useIsInverse } from '../../inverse';
+import styled from '@emotion/styled';
 
 /**
  * @children required