+++ packages/react-magma-dom/src/components/Select/Select.tsx
@@ -121,7 +121,7 @@ export function Select<T>(props: SelectProps<T>) {
   const toggleButtonProps = getToggleButtonProps({
     disabled: disabled,
     onBlur,
-    onKeyDown: event => {
+    onKeyDown: (event: any) => {
       if (
         getToggleButtonProps().id === document.activeElement.id &&
         (event.key === 'Enter' || event.key === ' ')
@@ -132,8 +132,8 @@ export function Select<T>(props: SelectProps<T>) {
 
       onKeyDown && typeof onKeyDown === 'function' && onKeyDown(event);
     },
-    onKeyPress,
-    onKeyUp,
+    onKeyPress: (event: any) => onKeyPress(event),
+    onKeyUp: (event: any) => onKeyUp(event),
     onFocus,
     ref,
   });