+++ packages/react-magma-dom/src/components/SkipLink/index.tsx
@@ -1,8 +1,8 @@
 import * as React from 'react';
-import styled from '../../theme/styled';
 import { ButtonColor, ButtonVariant } from '../Button';
 import { Hyperlink, HyperlinkProps } from '../Hyperlink';
 import { I18nContext } from '../../i18n';
+import styled from '@emotion/styled';
 
 export const TARGET_ID = 'reactMagmaMainContent';