+++ packages/react-magma-dom/src/components/Tabs/index.tsx
@@ -4,3 +4,5 @@ export * from './TabPanelsContainer';
 export * from './Tabs';
 export * from './TabsContainer';
 export * from './TabsScrollButtons';
+export * from './TabsScrollSpyContainer';
+export * from './TabScrollSpyPanel';