+++ packages/react-magma-dom/src/index.ts
@@ -48,6 +48,7 @@ export {
 export { Container, ContainerProps } from './components/Container';
 export { Datagrid, DatagridProps } from './components/Datagrid';
 export { DatePicker, DatePickerProps } from './components/DatePicker';
+export { inDateRange, getDateFromString } from './components/DatePicker/utils';
 export {
   Dropdown,
   DropdownAlignment,
@@ -206,6 +207,8 @@ export {
 export { Tab } from './components/Tabs/Tab';
 export { TabPanelsContainer } from './components/Tabs/TabPanelsContainer';
 export { TabPanel } from './components/Tabs/TabPanel';
+export { TabsScrollSpyContainer } from './components/Tabs/TabsScrollSpyContainer';
+export { TabScrollSpyPanel } from './components/Tabs/TabScrollSpyPanel';
 
 export { Tag, TagProps } from './components/Tag';
 
@@ -263,6 +266,7 @@ export * from './components/Tag';
 export * from './components/List';
 export * from './components/BlockQuote';
 export * from './components/ButtonGroup';
+export * from './components/TreeView';
 
 export * from './components/CharacterCounter';
 export * from './components/ToggleButton';