+++ website/react-magma-docs/gatsby-node.js
@@ -11,6 +11,9 @@ exports.onCreateWebpackConfig = ({ actions, plugins }) => {
       alias: {
         path: require.resolve('path-browserify'),
       },
+      fallback: {
+        'object.assign/polyfill': require.resolve("object.assign/polyfill.js"),
+      }
     },
     module: {
       rules: [